Finish mopping up the remains of the old Wikimedia API file
