+++ src/flickypedia/apis/wikimedia/__init__.py
@@ -1,4 +1,3 @@
-from .api import WikimediaApi
 from .exceptions import (
     WikimediaApiException,
     UnknownWikimediaApiException,
@@ -9,6 +8,29 @@ from .exceptions import (
 )
 from .language_methods import top_n_languages, LanguageMatch
 from .url_parser import get_filename_from_url
+
+from .base import HttpxImplementation
+
+from .category_methods import CategoryMethods
+from .language_methods import LanguageMethods
+from .structured_data_methods import StructuredDataMethods
+from .validator_methods import ValidatorMethods
+from .wikitext_methods import WikitextMethods
+from .upload_methods import UploadMethods
+from .user_methods import UserMethods
+
+
+class WikimediaApi(
+    HttpxImplementation,
+    CategoryMethods,
+    LanguageMethods,
+    StructuredDataMethods,
+    ValidatorMethods,
+    WikitextMethods,
+    UploadMethods,
+    UserMethods,
+):
+    pass
 
 
 __all__ = [
